Switch to single mach_hare impl.
