--- lib/logstash/outputs/rabbitmq/hot_bunnies.rb
@@ -1,1 +1,0 @@
-require "logstash/outputs/rabbitmq/march_hare"
